# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Music Distribution CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /releases:
    get:
      operationId: listReleases
      tags: [releases]
      summary: List all releases
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, pending_review, approved, delivering, delivered, live, error, takedown_requested, removed]
        - name: artist_id
          in: query
          schema:
            type: integer
        - name: label_id
          in: query
          schema:
            type: integer
        - name: release_type
          in: query
          schema:
            type: string
            enum: [single, album, ep, compilation]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of releases
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedReleases"
    post:
      operationId: createRelease
      tags: [releases]
      summary: Create a new release
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReleaseBody"
      responses:
        "201":
          description: Release created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Release"
  /releases/{id}:
    get:
      operationId: getRelease
      tags: [releases]
      summary: Get release by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Release details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReleaseDetail"
    put:
      operationId: updateRelease
      tags: [releases]
      summary: Update release
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReleaseBody"
      responses:
        "200":
          description: Release updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Release"
    delete:
      operationId: deleteRelease
      tags: [releases]
      summary: Delete release
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Release deleted
  /releases/{id}/status:
    patch:
      operationId: updateReleaseStatus
      tags: [releases]
      summary: Update release status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateReleaseStatusBody"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Release"
  /releases/import-upc:
    post:
      operationId: importReleaseByUpc
      tags: [releases]
      summary: Import release metadata by UPC
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportByUpcBody"
      responses:
        "200":
          description: Release metadata fetched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Release"

  # ==================== TRACKS ====================
...
  /splits:
    get:
      operationId: listSplits
      tags: [splits]
      summary: List revenue splits
      parameters:
        - name: release_id
          in: query
          schema:
            type: integer
        - name: track_id
          in: query
          schema:
            type: integer
        - name: artist_id
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of splits
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedSplits"
    post:
      operationId: createSplit
      tags: [splits]
      summary: Create revenue split
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSplitBody"
      responses:
        "201":
          description: Split created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Split"
  /splits/{id}:
    get:
      operationId: getSplit
      tags: [splits]
      summary: Get split by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Split details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Split"
    put:
      operationId: updateSplit
      tags: [splits]
      summary: Update split
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSplitBody"
      responses:
        "200":
          description: Split updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Split"
    delete:
      operationId: deleteSplit
      tags: [splits]
      summary: Delete split
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Split deleted

  # ==================== PAYOUTS ====================
...
components:
  schemas:
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
      required: [page, limit, total, totalPages]

    # ==================== DASHBOARD ====================
...
    Release:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        releaseType:
          type: string
          enum: [single, album, ep, compilation]
        status:
          type: string
          enum: [draft, pending_review, approved, delivering, delivered, live, error, takedown_requested, removed]
        upc:
          type: string
          nullable: true
        artistId:
          type: integer
        artistName:
          type: string
        labelId:
          type: integer
          nullable: true
        labelName:
          type: string
          nullable: true
        coverUrl:
          type: string
          nullable: true
        genre:
          type: string
          nullable: true
        releaseDate:
          type: string
          nullable: true
        language:
          type: string
          nullable: true
        isExplicit:
          type: boolean
        territories:
          type: array
          items:
            type: string
        totalTracks:
          type: integer
        pLine:
          type: string
          nullable: true
        cLine:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, title, releaseType, status, artistId, artistName, isExplicit, territories, totalTracks, createdAt, updatedAt]
    ReleaseDetail:
      allOf:
        - $ref: "#/components/schemas/Release"
        - type: object
          properties:
            tracks:
              type: array
              items:
                $ref: "#/components/schemas/Track"
    CreateReleaseBody:
      type: object
      properties:
        title:
          type: string
        releaseType:
          type: string
          enum: [single, album, ep, compilation]
        artistId:
          type: integer
        labelId:
          type: integer
          nullable: true
        upc:
          type: string
          nullable: true
        coverUrl:
          type: string
          nullable: true
        genre:
          type: string
          nullable: true
        releaseDate:
          type: string
          nullable: true
        language:
          type: string
          nullable: true
        isExplicit:
          type: boolean
          default: false
        territories:
          type: array
          items:
            type: string
          default: ["WW"]
        pLine:
          type: string
          nullable: true
        cLine:
          type: string
          nullable: true
      required: [title, releaseType, artistId]
    UpdateReleaseStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [draft, pending_review, approved, delivering, delivered, live, error, takedown_requested, removed]
        note:
          type: string
          nullable: true
      required: [status]
    ImportByUpcBody:
      type: object
      properties:
        upc:
          type: string
        source:
          type: string
          enum: [spotify, apple, musicbrainz]
          default: spotify
      required: [upc]
    PaginatedReleases:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Release"
        pagination:
          $ref: "#/components/schemas/Pagination"
      required: [data, pagination]

    # ==================== TRACKS ====================
    Track:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        isrc:
          type: string
          nullable: true
        releaseId:
          type: integer
          nullable: true
        releaseName:
          type: string
          nullable: true
        artistId:
          type: integer
        artistName:
          type: string
        trackNumber:
          type: integer
          nullable: true
        durationSeconds:
          type: integer
          nullable: true
        genre:
          type: string
          nullable: true
        language:
          type: string
          nullable: true
        isExplicit:
          type: boolean
        composerName:
          type: string
          nullable: true
        lyricistName:
          type: string
          nullable: true
        iswc:
          type: string
          nullable: true
        audioUrl:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, title, artistId, artistName, isExplicit, createdAt, updatedAt]
...
    SplitParticipant:
      type: object
      properties:
        entityType:
          type: string
          enum: [artist, label, distributor, producer, author, custom]
        entityId:
          type: integer
          nullable: true
        entityName:
          type: string
        percentage:
          type: number
      required: [entityType, entityName, percentage]
    Split:
      type: object
      properties:
        id:
          type: integer
        releaseId:
          type: integer
          nullable: true
        releaseName:
          type: string
          nullable: true
        trackId:
          type: integer
          nullable: true
        trackName:
          type: string
          nullable: true
        participants:
          type: array
          items:
            $ref: "#/components/schemas/SplitParticipant"
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, participants, createdAt, updatedAt]
    CreateSplitBody:
      type: object
      properties:
        releaseId:
          type: integer
          nullable: true
        trackId:
          type: integer
          nullable: true
        participants:
          type: array
          items:
            $ref: "#/components/schemas/SplitParticipant"
      required: [participants]
    PaginatedSplits:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Split"
        pagination:
          $ref: "#/components/schemas/Pagination"
      required: [data, pagination]

    # ==================== PAYOUTS ====================
...